Revert to kernel version 3.10.65

+++ drivers/misc/mediatek/mach/mt6752/include/mach/hw_watchpoint.h
@@ -1,12 +1,12 @@
 #ifndef __HW_BREAKPOINT_H
 #define __HW_BREAKPOINT_H
 
-typedef int (*wp_handler)(unsigned long addr);
+typedef int (*wp_handler)(unsigned int addr);
 
 struct wp_event
 {
-    unsigned long virt;
-    unsigned long phys;
+    unsigned int virt;
+    unsigned int phys;
     int type;
     wp_handler handler;
     int in_use;
